WEEK 3/May 20/requirements.txt
google-generativeai==0.8.5
autogen-core==0.5.7
autogen-agentchat==0.5.7
